Stop modifying a process’s command when the droplet is set
We only want process.command to be reflect the user's explicit wishes.
We want to favor the manifest/API over the detected start command. Part
of that hierarchy will be to only store those explicit wishes in
process.command and to fall back to the detected start command on the
droplet.

[#157756113]

Signed-off-by: Mike Kenyon <mkenyon@pivotal.io>